feat: Add matrix multiplication algorithm
The code changes introduce a new function `mult` in the `E3.py` file. This function performs matrix multiplication by dividing each matrix into four sub-matrices of dimension n/2. It iterates through the sub-matrices and calculates the corresponding elements of the resulting matrix. The resulting matrix is then returned.

This commit message suggests that the purpose of the code changes is to add a matrix multiplication algorithm to the project.